+++ lib/fastlane/plugin/emerge/helper/git.rb
@@ -1,20 +1,27 @@
+require 'fastlane_core/print_table'
 require 'open3'
 
 module Fastlane
   module Helper
     module Git
       def self.branch
-        stdout, _, status = Open3.capture3("git rev-parse --abbrev-ref HEAD")
+        shell_command = "git rev-parse --abbrev-ref HEAD"
+        UI.command(shell_command)
+        stdout, _, status = Open3.capture3(shell_command)
         stdout.strip if status.success?
       end
 
       def self.sha
-        stdout, _, status = Open3.capture3("git rev-parse HEAD")
+        shell_command = "git rev-parse HEAD"
+        UI.command(shell_command)
+        stdout, _, status = Open3.capture3(shell_command)
         stdout.strip if status.success?
       end
 
       def self.base_sha
-        stdout, _, status = Open3.capture3("git merge-base #{remote_head_branch} #{branch}")
+        shell_command = "git merge-base #{remote_head_branch} #{branch}"
+        UI.command(shell_command)
+        stdout, _, status = Open3.capture3(shell_command)
         return nil if stdout.strip.empty? || !status.success?
         current_sha = sha
         stdout.strip == current_sha ? nil : stdout.strip
@@ -28,7 +35,9 @@ module Fastlane
 
       def self.remote_head_branch(remote = primary_remote)
         return nil if remote.nil?
-        stdout, _, status = Open3.capture3("git remote show #{remote}")
+        shell_command = "git remote show #{remote}"
+        UI.command(shell_command)
+        stdout, _, status = Open3.capture3(shell_command)
         return nil if stdout.nil? || !status.success?
         stdout
           .split("\n")
@@ -40,12 +49,16 @@ module Fastlane
 
       def self.remote_url(remote = primary_remote)
         return nil if remote.nil?
-        stdout, _, status = Open3.capture3("git config --get remote.#{remote}.url")
+        shell_command = "git config --get remote.#{remote}.url"
+        UI.command(shell_command)
+        stdout, _, status = Open3.capture3(shell_command)
         stdout if status.success?
       end
 
       def self.remote
-        stdout, _, status = Open3.capture3("git remote")
+        shell_command = "git remote"
+        UI.command(shell_command)
+        stdout, _, status = Open3.capture3(shell_command)
         stdout.split("\n") if status.success?
       end
     end
